\name{rpcap-package}
\alias{rpcap-package}
\alias{rpcap}
\docType{package}
\title{
 Package rpcap
}
\description{
	A Package to process network packet header data.
}
\details{

The package rpcap contains example traces of network traffic extracted from the DARPA Intrusion Detection Evaluation Data Set (see: packetHeader),
a two list which map protocol number to protocol name (see: protocol) as well as port number to name (see: port) extracted from IANA.  
The provided functions aim to easily generate statistics to interpret and analyze packet header data in csv format.

}

\author{
Frankfurt University of Applied Sciences
Maintainer: Denis Hock
}

\references{
Hock, D., and M. Kappes. "Using R for Anomaly Detection in Network Traffic." Fifth International Conference on Internet Technologies & Applications. 2013.
}

\keyword{ package }

\examples{
}
